Initial Port from `ember-cli-rails`
Extracts `include_ember_script_tags` and `include_ember_stylesheet_tags`
from `ember-cli-rails` proper, including classes focussed on finding and
inlining EmberCLI-generated assets.

The gem's version is set to `0.6.1` to match the version of
`ember-cli-rails` in which it was added as a direct dependency.

In future versions of `ember-cli-rails` (i.e. `0.7.0` and up), when
asset management will be removed, this gem will be removed as a direct
dependency, and will exist as an optional addon.