fix: Update asset_map to handle assets more than one level deep
When a CSS files is used in the `index.html` that has  a path like such: `/assets/dir1/dir2/dir3/library.css`. The asset map throws an exception that it can't be found. This is because the DirectoryAssetMap only looks at children, and not the entire tree. This now returns all the files found recursively for it to check. In addition, it returns the relative path, so there isn't a risk of two files in different directories colliding.